requirements-dev.txt
configparser==3.5.1
enum34==1.1.6
flake8==3.6.0
mccabe==0.6.1
pycodestyle==2.4.0
pyflakes==2.0.0
PyYAML==3.13
